to stable

+++ core/class/docker2.class.php
@@ -16,21 +16,22 @@
  * along with Jeedom. If not, see <http://www.gnu.org/licenses/>.
  */
 
-/* * ***************************Includes********************************* */
 require_once __DIR__  . '/../../../../core/php/core.inc.php';
 
 class docker2 extends eqLogic {
-   /*     * *************************Attributs****************************** */
-
-
-   /*     * ***********************Methode static*************************** */
 
    public static function execCmd($_cmd, $_docker_number = 1, $_format = "{{json . }}", $_notrunc = true) {
+      if ($_cmd == '') {
+         return '';
+      }
       if ($_format != '') {
          $_cmd .= ' --format "' . $_format . '"';
          if ($_notrunc != '') {
             $_cmd .= ' --no-trunc';
          }
+      }
+      if (strpos($_cmd, ' 2>&1') === false) {
+         $_cmd .= ' 2>&1';
       }
       $config = config::byKey('docker_config_' . $_docker_number, 'docker2');
       if ($config['mode'] == 'local') {
@@ -38,16 +39,16 @@ class docker2 extends eqLogic {
          $retval = null;
          exec($_cmd, $output, $retval);
          if ($retval != 0) {
-            throw new Exception(__('Erreur d\'éxécution de la commande : ', __FILE__) . $_cmd . ' (' . $retval . ') => ' . json_encode($output));
+            throw new Exception(__("Erreur d'exécution de la commande", __FILE__) . ' : ' . $_cmd . ' (' . $retval . ') => ' . json_encode($output));
          }
       } else if ($config['mode'] == 'ssh') {
          $connection = ssh2_connect($config['ip'], $config['port']);
          if ($connection === false) {
-            throw new Exception(__('Impossible de se connecter sur :', __FILE__) . ' ' . $config['ip'] . ':' . $config['mode']);
+            throw new Exception(__('Impossible de se connecter sur', __FILE__) . ' : ' . $config['ip'] . ':' . $config['mode']);
          }
          $auth = @ssh2_auth_password($connection, $config['username'], $config['password']);
          if (false === $auth) {
-            throw new Exception(__('Echec de l\'authentification SSH', __FILE__));
+            throw new Exception(__("Echec de l'authentification SSH", __FILE__));
          }
          $stream = ssh2_exec($connection, $_cmd);
          stream_set_blocking($stream, true);
@@ -75,7 +76,7 @@ class docker2 extends eqLogic {
 
    public static function restoreDockerFile($_filepath) {
       if (!file_exists($_filepath)) {
-         throw new Exception(__('Fichier de backup non trouvé : ', __FILE__) . $_filepath);
+         throw new Exception(__('Fichier de sauvegarde non trouvé', __FILE__) . ' : ' . $_filepath);
       }
       shell_exec('sudo tar -xf ' . $_filepath . ' -C /');
    }
@@ -92,7 +93,7 @@ class docker2 extends eqLogic {
                self::pull($i);
             }
          } catch (Exception $exc) {
-            log::add('docker2', 'error', __('Expression cron non valide pour docker ', __FILE__) . $config['cron']);
+            log::add('docker2', 'error', __('Expression cron non valide pour Docker', __FILE__) . ' ' . $i . ' : ' . $config['cron']);
          }
       }
    }
@@ -232,8 +233,6 @@ class docker2 extends eqLogic {
       return $return;
    }
 
-   /*     * *********************Méthodes d'instance************************* */
-
    public function postSave() {
       $cmd = $this->getCmd(null, 'start');
       if (!is_object($cmd)) {
@@ -251,7 +250,7 @@ class docker2 extends eqLogic {
       if (!is_object($cmd)) {
          $cmd = new docker2Cmd();
          $cmd->setLogicalId('stop');
-         $cmd->setName(__('Arreter', __FILE__));
+         $cmd->setName(__('Arrêter', __FILE__));
       }
       $cmd->setDisplay('icon', '<i class="fas fa-stop"></i>');
       $cmd->setType('action');
@@ -405,8 +404,25 @@ class docker2 extends eqLogic {
          if ($this->getIsEnable() == 1 && $this->getLogicalId() == '') {
             $this->create();
          }
+
+         $cmd = $this->getCmd(null, 'update');
+         if (!is_object($cmd)) {
+            $cmd = new docker2Cmd();
+            $cmd->setLogicalId('update');
+            $cmd->setName(__('Mettre à jour', __FILE__));
+         }
+         $cmd->setDisplay('icon', '<i class="fas fa-spinner"></i>');
+         $cmd->setType('action');
+         $cmd->setSubType('other');
+         $cmd->setEqLogic_id($this->getId());
+         $cmd->save();
       } else {
          $cmd = $this->getCmd(null, 'receate');
+         if (is_object($cmd)) {
+            $cmd->remove();
+         }
+
+         $cmd = $this->getCmd(null, 'update');
          if (is_object($cmd)) {
             $cmd->remove();
          }
@@ -429,18 +445,18 @@ class docker2 extends eqLogic {
       switch ($this->getConfiguration('create::mode')) {
          case 'jeedom_run':
             if ($this->getConfiguration('create::run') == '') {
-               throw new Exception(__('Vous ne pouvez lancer la création d\'un docker depuis Jeedom sans commande de création', __FILE__));
+               throw new Exception(__('Impossible de créer un conteneur depuis Jeedom sans commande de création', __FILE__));
             }
             $cmd = system::getCmdSudo() . ' docker run ';
             $cmd .= ' ' . str_replace('docker run', '', str_replace(system::getCmdSudo() . ' docker run', '', $this->getConfiguration('create::run')));
             if (strpos($cmd, ' -d') === false) {
-               throw new Exception(__('Il n\'est pas possible de lancer un contenaire depuis Jeedom sans l\'option deamon', __FILE__));
+               throw new Exception(__('Impossible de démarrer un conteneur depuis Jeedom sans option de démon', __FILE__));
             }
             self::execCmd($cmd, $this->getConfiguration('docker_number', 1), null);
             break;
          case 'jeedom_compose':
             if ($this->getConfiguration('create::compose') == '') {
-               throw new Exception(__('Vous ne pouvez lancer la création d\'un docker depuis Jeedom sans docker compose', __FILE__));
+               throw new Exception(__('Impossible de créer un conteneur depuis Jeedom sans Docker Compose', __FILE__));
             }
             $filename = '/tmp/' . bin2hex(random_bytes(10)) . '.yml';
             file_put_contents($filename, $this->getConfiguration('create::compose'));
@@ -448,7 +464,7 @@ class docker2 extends eqLogic {
             unlink($filename);
             break;
          default:
-            throw new Exception(__('La création de ce docker n\'est pas gérée par Jeedom', __FILE__));
+            throw new Exception(__('Impossible de créer ce conteneur depuis Jeedom', __FILE__));
             break;
       }
 
@@ -459,22 +475,56 @@ class docker2 extends eqLogic {
    }
 
    public function rm() {
+      if ($this->getConfiguration('id') == '') {
+         return;
+      }
       self::execCmd(system::getCmdSudo() . ' docker rm -f ' . $this->getConfiguration('id'), $this->getConfiguration('docker_number'), null);
    }
 
    public function logs() {
+      if ($this->getConfiguration('id') == '') {
+         return;
+      }
       return self::execCmd(system::getCmdSudo() . ' docker logs -t -n 100 ' . $this->getConfiguration('id') . ' 2>&1', $this->getConfiguration('docker_number'), null);
    }
 
    public function inspect() {
+      if ($this->getConfiguration('id') == '') {
+         return;
+      }
       return self::execCmd(system::getCmdSudo() . ' docker inspect ' . $this->getConfiguration('id'), $this->getConfiguration('docker_number'), '{{json . }}', false);
+   }
+
+   public function updateImage() {
+      if ($this->getConfiguration('id') == '') {
+         return;
+      }
+      if (!in_array($this->getConfiguration('create::mode'), array('jeedom_run', 'jeedom_compose'))) {
+         throw new Exception(__("Impossible de mettre à jour l'image de ce conteneur depuis Jeedom.", __FILE__));
+      }
+      $inspect = $this->inspect();
+      try {
+         $this->stopDocker();
+         sleep(5);
+      } catch (\Throwable $th) {
+      }
+      try {
+         $this->rm();
+         sleep(5);
+      } catch (\Throwable $th) {
+      }
+      self::execCmd(system::getCmdSudo() . ' docker rmi ' . $inspect[0]['Config']['Image'], $this->getConfiguration('docker_number'), '');
+      sleep(5);
+      $this->create();
+      sleep(5);
+      $this->startDocker();
    }
 
    public function backupDocker() {
       if ($this->getConfiguration('saveMount') == 0) {
-         throw new Exception(__('Ce docker n\'est pas sauvegardé', __FILE__));
+         throw new Exception(__("Ce conteneur n'est pas sauvegardé", __FILE__));
       }
-      $folder = __DIR__ . '/../../data/backup';
+      $folder = __DIR__ . '/../../data/mountsBackup';
       if (!file_exists($folder)) {
          mkdir($folder);
       }
@@ -497,11 +547,11 @@ class docker2 extends eqLogic {
 
    public function restoreDocker() {
       if ($this->getConfiguration('saveMount') == 0) {
-         throw new Exception(__('Ce docker n\'est pas sauvegarder', __FILE__));
+         throw new Exception(__("Ce Docker n'est pas sauvegardé", __FILE__));
       }
-      $filepath = __DIR__ . '/../../data/backup/' . $this->getId() . '.tar.gz';
+      $filepath = __DIR__ . '/../../data/mountsBackup/' . $this->getId() . '.tar.gz';
       if (!file_exists($filepath)) {
-         throw new Exception(__('Aucune sauvegarde trouvée pour ce docker', __FILE__));
+         throw new Exception(__('Aucune sauvegarde trouvée pour ce conteneur', __FILE__));
       }
       self::restoreDockerFile($filepath);
       $this->restartDocker();
@@ -512,7 +562,7 @@ class docker2 extends eqLogic {
       if (isset($template['docker-compose'])) {
          $docker_compose = file_get_contents(__DIR__ . '/../config/template/' . $template['docker-compose']['file']);
          if ($docker_compose == '') {
-            throw new Exception(__('Erreur lors de la récuperation du docker compose : ', __FILE__) . $template['docker-compose']['file']);
+            throw new Exception(__('Erreur lors de la récupération du Docker Compose', __FILE__) . ' : ' . $template['docker-compose']['file']);
          }
       }
       $this->import($template['eqLogic'], true);
@@ -545,18 +595,9 @@ class docker2 extends eqLogic {
          }
       }
    }
-
-   /*     * **********************Getteur Setteur*************************** */
 }
 
 class docker2Cmd extends cmd {
-   /*     * *************************Attributs****************************** */
-
-
-   /*     * ***********************Methode static*************************** */
-
-
-   /*     * *********************Methode d'instance************************* */
 
    public function formatValueWidget($_value) {
       if ($this->getLogicalId() != 'state') {
@@ -586,13 +627,8 @@ class docker2Cmd extends cmd {
       } else if ($this->getLogicalId() == 'remove') {
          $eqLogic->rm();
       } else if ($this->getLogicalId() == 'update') {
-         $eqLogic->rm();
-         sleep(2);
-         docker2::execCmd(system::getCmdSudo() . ' docker rmi ' . $eqLogic->getConfiguration('image') . ' 2>&1', $eqLogic->getConfiguration('docker_number'), null);
-         $eqLogic->create();
+         $eqLogic->updateImage();
       }
       docker2::pull();
    }
-
-   /*     * **********************Getteur Setteur*************************** */
 }
